Add new prop selectSingleDay to DateRangeInput component
By setting the selectSingleDay option to true (which is the default
value) the user can select a single day instead of a date range. If it
is set to false, a single day cannot be selected. You can select the
single day by clicking on the date twice (selected the start and end
date as the same date)

This commit also adds a yarn.lock file so that yarn can be used to
install dependencies.

Webpack has been upgraded to v2.3.0 and the loaders have been upgraded
as well. EsLint and some of its rules have been upgraded.