Updates React to v15.6.1 and other dependencies
Updates React and react-dom to the latest version and updates the
DateRangeInput component to use the prop-types package and converts it
to use es6 classes instead of React.createClass().

Updates the build script in package.json to use yarn instead of npm.

NOTE: Because the dependent package react-daterange-picker still has not
been updated to use React 15.5.x and up there are still warnings that
are shown in the console. The dependency will be updated as soon as
the needed updates are merged in.