Fix validation issues and React lifecycle issue
Fixes the issue in the `getDerivedStateFromProps` function where it was
attempting to reference `this` inside of a static function. Also, fixes
an issue where validation failed from passing in the default date ranges
specified in the DateRangeInput file, and the solution was to move them
inside of the React class and set it as a static property.