Fix issue when validating the dates
Fixes an issue that used the constructor.name of the dates used to check
if it is a valid Moment object. When the code is compiled for production
the name value is changed because it is minified.

Also, upgrades webpack and webpack-cli to the latest version and fixes a
config issue with the webpack.cvonfig.js file.