Fix issue when clear dates is called and moment-range isSame is used
There was an issue with setting the state `value` to null when the clear
dates function is called and then the moment-range isSame function is
called. It would throw an error and break everything because null was
getting passed in to be compared with an actual moment-range object.
Instead of setting the state `value` to null this sets it to a
moment-range object.