Fix issues with custom range picking
Previously the default value was set to the value returned from
`moment.range()`. This caused problems because it is a huge date range
that gets returned. This commit fixes some of the issues that were
caused when setting this value when a custom range was selected and
ensures that the `isSame` function doesn't get called on a null value as
well as trying to get the `start` value from a null value.