Fix bug when selecting one date and update ranges
- Fixes a bug that occurred when selecting a single day and then
clicking outside of the date picker causing it to close. The date was
selected but wasn't reflected in the input button where the range of the
date selected is displayed.
- Updates the endDate for the default ranges so that they use the
`endOf` moment function instead of `startOf` moment function.

NOTE: This commit also upgrades the project dependencies, to the latest
version.